feat(website): replace product Demo section with a register Signoff
The product page closed on a static AppPreview mockup ("Try it live").
Swap it for the shared Signoff closer — a focused register CTA, the same
component the customers / why-twenty / partners pages use — so the page
ends by driving sign-ups instead of re-showing the product.

- Add ProductSignoff: heading "Start moving faster today.", a supporting
  line, and the Get started + Talk to us pair (matching the canonical
  CustomersCatalogSignoff). Uses SITE_URLS.appWelcome per the site-urls
  rule rather than inlining the URL.
- Add an opt-in headingMaxWidth prop to Signoff so the heading wraps to
  two lines by width (text-wrap: balance) instead of a hardcoded \n. The
  existing signoffs don't pass it, so they are untouched.
- Remove the now-unused product-demo section and its 192KB background.
