+++ packages/twenty-website/src/app/[locale]/(site)/product/page.tsx
@@ -12,9 +12,9 @@ import {
 } from '@/platform/seo';
 import { Faq } from '@/sections/faq';
 import { Menu } from '@/sections/menu';
-import { ProductDemo } from '@/sections/product-demo';
 import { ProductFeature } from '@/sections/product-feature';
 import { ProductHero } from '@/sections/product-hero';
+import { ProductSignoff } from '@/sections/product-signoff';
 import { ProductStepper } from '@/sections/product-stepper';
 import { ProductThreeCards } from '@/sections/three-cards';
 import { TrustedBy } from '@/sections/trusted-by';
@@ -51,7 +51,7 @@ export default async function ProductPage({
           <ProductFeature />
           <ProductThreeCards />
           <ProductStepper />
-          <ProductDemo />
+          <ProductSignoff />
           <Faq />
         </main>
       </MenuStyleProvider>
