--- packages/twenty-website/src/sections/product-demo/index.ts
@@ -1,1 +1,0 @@
-export { ProductDemo } from './ProductDemo';
