+++ packages/twenty-website/src/sections/product-signoff/index.ts
@@ -1,0 +1,1 @@
+export { ProductSignoff } from './ProductSignoff';
